fix: long-standing bug with windows not appearing
With specific scripts & platforms (Linux & macOS), some pop-up windows
do not appear on the screen when scripts are initially selected.

Instead, they are drawn on screen with a window size of 0, and users
have to hunt around the screen to find the window.

Countless seconds of searching online did not find a result, until this
particular [one][0] was stumbled on.

tl;dr - in the future, ensure that any windowing is done as per below:

  frame.pack();
  frame.setLocationRelativeTo(null);
  frame.setVisible();
  frame.requestFocusInWindow();

This ensures that all windows are sized correctly.

[0]: https://stackoverflow.com/questions/8193801/how-to-set-specific-window-frame-size-in-java-swing