New Scripts NoBankSuperheat and WhiteUnicorns
NoBankSuperheat mines iron in khazard and uses superheat to make iron bars (100% completion chance) and then makes iron plates at the nearby anvil. Only resource required is nature runes. This method is designed for UIM smithing/magic training and is fairly efficient.

WhiteUnicorns - kills the 2 white unicorns north west of catherby bank. This is extremely inefficient but meant as an alternative gathering method for hardcore accounts not wanting to visit the wilderness for unicorn horns.