Add mining and noting ess option to damRc
This will likely be for UIM runecrafting training in the future. Process to retrieve rune stones will be go to any bank, pay 1k gp to teleport to varrock center with auctioneer, talk to mort to unnote ess. Then probably cape teleport or magic teleport somewhere closer and walk to alter.