Add sleep assets to jar and extract when missing
This will solve missing assets on release tasks or when running the jar solo without any other resources in the dir. Ideally Idle jar is self sufficient and can extract any assets it needs on the fly without worrying about missing dependencies if the jar is moved or something is deleted.